% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{a data frame of training set cases.}

\item{cl}{a matrix or data frame of test set cases.}

\item{k_nn}{an integer representing the number of neighbors considered.}

\item{k_cv}{an integer representing the number of folds.}
}
\value{
A list containing the following component:
\describe{
\item{\code{class}}{Factor of classifications of test set.}
\item{\code{cv_err}}{cross validation error}
}
}
\description{
This function performs cross validation on k-Nearest Neighbors classification.
More information about k-Nearest Neighbors classification are available at
\code{\link[class]{knn}}.
}
\examples{
train <- data.frame(rbind(iris3[1:25,,1], iris3[1:25,,2], iris3[1:25,,3]))
cl <- factor(c(rep("s",25), rep("c",25), rep("v",25)))
my_knn_cv(train, cl, k_nn = 5, k_cv = 5)

}
\keyword{prediction}
