% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My lm}
\usage{
my_lm(formula, data.frame)
}
\arguments{
\item{formula}{An object of class "formula". A symbolic description of the
model to be fitted.}

\item{data.frame}{A data frame containing the variables in the model.}
}
\value{
A table with rows for each coefficient (including the intercept) and
columns for the estimate, standard error, t-value, and p-value.
}
\description{
My_lm is used to fit linear models.
}
\examples{
my_lm(my_penguins$bill_length_mm, my_penguins$bill_depth_mm)
my_lm(my_penguins$body_mass_g, my_penguins$flipper_length_mm)

}
\keyword{linear}
\keyword{lm}
\keyword{model}
