% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{an integer representing the number of folds.}
}
\value{
Numeric representing the cross-validation error, which is the average
MSE across all \code{k} folds.
}
\description{
This function performs cross validation on random forest regression to predict
output class \code{body_mass_g} using covariates \code{bill_length_mm},
\code{bill_depth_mm}, \code{flipper_length_mm} within \code{\link[palmerpenguins]{penguins}}
dataset. More information about random forest regression are available at
\code{\link[randomForest]{randomForest}}.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
