% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{My t-test}
\usage{
my_t.test(x, alternative = c("two.sided", "less", "greater"), mu = 0)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.
This only accepts "two.sided", "less", or "greater".}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list of the following elements: test statistic, degrees of freedom,
alternative, and p-value.
}
\description{
Performs a one sample t-test on vectors of data.
}
\examples{
my_t.test(my_penguins$body_mass_g, alternative = "two.sided", mu = 0)
my_t.test(my_penguins$bill_length_mm, alternative = "greater", mu = 1)

}
\keyword{statistic}
\keyword{t-test}
\keyword{test}
