% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" that gives a symbolic description
of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}
}
\value{
A table containing the following componenrmt:
\describe{
\item{\code{Estimate}}{the linear regression coefficients of each variables.}
\item{\code{Std. Error}}{the standard error for coefficients.}
\item{\code{t value}}{the value of the t_statistic for coefficients}
\item{\code{Pr(>|t|)}}{P value for tests for coefficients}
}
}
\description{
This function fits a linear model.
}
\examples{
my_lm(mpg ~ wt, data = mtcars)

}
\keyword{inference}
\keyword{prediction}
