% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{Student's t-test}
\usage{
my_t.test(x, alternative = "two.sided", mu)
}
\arguments{
\item{x}{a non-empty numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis.
This only accept \code{"two.sided"} (default), \code{"greater"}, or \code{"less"}.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
A list containing the following component:
\describe{
\item{\code{test_stat}}{the value of the t_statistic.}
\item{\code{df}}{the degree of freedom for the t-statistic.}
\item{\code{alternative}}{a character string describing the alternative hypothesis.}
\item{\code{p_val}}{the p-value for the test.}
}
}
\description{
Performs one sample t-test on vectors of data.
}
\examples{
x <- rnorm(20, mean = 0, sd = 1)
my_t.test(x, mu = 0)
my_t.test(x, alternative = "less", mu = 0)

}
\keyword{inference}
